+++ src/util.rs
@@ -10,7 +10,6 @@ pub fn width_of_text(text: &str, font: &Font, scale: Scale) -> Pt {
                 .get_data()
                 .map(|data| data.unit_h_metrics.advance_width as f64)
                 .unwrap()
-        })
-        .sum();
+        }).sum();
     Pt(glyph_space_width * scale.x as f64 / units_per_em)
 }